// src/models/vehicle.interface.ts
export interface Vehicle {
    uid?: string;
    model: string;
    regdate: string;
    regno: string;
    chassisno: string;
    passengers: number;
    doors: number;
    bags: number;
    cartype: string;
    aircondition: string;
    status: string;
    price: number;
    image: string;
    category?: string;
}